Remove non-PLUGIN_SERVER_INGESTION code
